seperated dom and fixed styles
